🎨 feature 폴더를 정리한다
