+++ src/features/export/ui/export-button.tsx
@@ -1,21 +1,9 @@
 import { Download } from 'lucide-react';
-import { useRecordingStore } from '@/entities/record';
 import { Button } from '@/shared/ui';
+import { useExportEvents } from '../hooks/use-export-events';
 
 const ExportButton = () => {
-  const { groupedEvents } = useRecordingStore();
-
-  const handleExport = () => {
-    const blob = new Blob([JSON.stringify(groupedEvents, null, 2)], {
-      type: 'application/json',
-    });
-    const url = URL.createObjectURL(blob);
-    const a = document.createElement('a');
-    a.href = url;
-    a.download = `api-events-${new Date().toISOString()}.json`;
-    a.click();
-    URL.revokeObjectURL(url);
-  };
+  const { handleExport } = useExportEvents();
 
   return (
     <Button variant="ghost" size="icon" onClick={handleExport}>
