+++ src/features/switch-ui-mode/css/open-devtools-button.css.ts
